Give user membership in the outside /dev/input group at build and runtime (#12)
This solves the problem raised in
https://github.com/marcindulak/stt-mcp-server-linux/pull/8#pullrequestreview-3264626994

The /dev/input group owner ID may differ outside/inside the container.
For the keyboard detection to work inside of the container, the user
inside of the container must be the member of the /dev/input group ID
present outside of the container.

The group id of the owner of `/dev/input` outside of the container needs
to be determined dynamically, since it may be very different from the
usual 900-range.

```
17:01 $ getent group input | cut -d: -f3
107
```

```
17:01 $ docker run -it stt-mcp-server-linux:latest bash -c "getent group input | cut -d: -f3"
997
```